Working on the edit component
