+++ imports/ui/components/editTaskForm/editTaskForm.html
@@ -1,0 +1,3 @@
+<template name="editTaskForm">
+  {{> taskForm }}
+</template>
