+++ imports/ui/components/editTaskForm/editTaskForm.js
@@ -1,0 +1,4 @@
+import { Template } from 'meteor/templating';
+
+import './editTaskForm.html';
+import '../taskForm/taskForm.js';
