+++ imports/ui/components/newTaskForm/newTaskForm.html
@@ -1,0 +1,3 @@
+<template name="newTaskForm">
+  {{> taskForm }}
+</template>
