+++ imports/ui/components/newTaskForm/newTaskForm.js
@@ -1,0 +1,4 @@
+import { Template } from 'meteor/templating';
+
+import './newTaskForm.html';
+import '../taskForm/taskForm.js';
