--- imports/ui/components/newTaskModal/newTaskModal.html
@@ -1,18 +1,0 @@
-<template name="newTaskModal">
-  <button class="btn btn-primary btn-lg btn-block btn-xl" data-toggle="modal" data-target=".bd-example-modal-lg">
-    Add Task
-  </button>
-
-
-  <div class="modal fade bd-example-modal-lg" tabindex="-1" role="dialog" aria-labelledby="myLargeModalLabel" aria-hidden="true">
-    <div class="modal-dialog modal-lg">
-
-      <div class="modal-content">
-        <div class = "container">
-          {{> newTask}}
-        </div>
-      </div>
-
-    </div>
-  </div>
-</template>
