--- imports/ui/components/newTaskModal/newTaskModal.js
@@ -1,4 +1,0 @@
-import { Template } from 'meteor/templating';
-
-import './newTaskModal.html';
-import '../newTask/newTask.js';
