+++ imports/ui/components/taskForm/taskForm.html
@@ -1,4 +1,4 @@
-<template name="newTask">
+<template name="taskForm">
   <h2>Add Task</h2>
     <form class="new-task">
       <div class="form-group">
